requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.5
setuptools>=58.0.0
psycopg2-binary==2.9.10
